requirements.txt
 # requirements.txt - 外汇周报 AI Agent
# 安装命令：pip install -r requirements.txt
# 如果 akshare 安装失败，请尝试：
#   pip install akshare -i http://mirrors.aliyun.com/pypi/simple/ --trusted-host=mirrors.aliyun.com

# Web 框架
streamlit

# LLM API
openai>=1.0.0

# 环境变量
python-dotenv

# 数据采集（核心）
akshare          # 国内金融数据接口（免费，不需要 API Key）
fredapi          # FRED 美国宏观数据（包含DXY美元指数）
yfinance         # Yahoo Finance - DXY美元指数备选数据源
requests
requests[socks]  # 支持 SOCKS5 代理

# 数据处理
pandas
numpy

# 网页解析
beautifulsoup4
lxml
